+++ src/components/charts/charts.js
@@ -447,6 +447,8 @@ charts.addLegend = function (series, chartType, settings, container) {
  * @param {object} settings [description]
  */
 charts.handleElementClick = function (line, series, settings) {
+  const api = $(settings?.svg?.node()).closest('.chart-container').data('chart');
+  const noTrigger = api?.initialSelectCall;
   const idx = $(line).index();
   const elem = series[idx];
   let selector;
@@ -457,13 +459,13 @@ charts.handleElementClick = function (line, series, settings) {
 
   if (settings.type === 'pie' || settings.type === 'donut') {
     selector = d3.select(settings.svg.selectAll('.slice').nodes()[idx]);
-  } else if (settings.type === 'column-positive-negative') {
+  } else if (/positive-negative/.test(settings.type)) {
     if (!elem.option || (elem.option && elem.option === 'target')) {
       selector = settings.svg.select('.target-bar');
     } else {
       selector = settings.svg.select(`.bar.${elem.option}`);
     }
-  } else if (['column', 'bar', 'bar-stacked', 'bar-grouped', 'bar-normalized', 'column-grouped', 'column-stacked', 'column-positive-negative'].indexOf(settings.type) !== -1) {
+  } else if (['column', 'bar', 'bar-stacked', 'bar-grouped', 'bar-normalized', 'column-grouped', 'column-stacked', 'column-positive-negative', 'positive-negative'].indexOf(settings.type) !== -1) {
     // Grouped or singlular
     if (settings.isGrouped || settings.isSingle) {
       selector = settings.svg.select(`.series-${idx}`);
@@ -475,18 +477,46 @@ charts.handleElementClick = function (line, series, settings) {
   }
 
   if (['radar', 'pie', 'donut', 'column', 'bar', 'bar-stacked', 'bar-grouped', 'bar-normalized',
-    'column-grouped', 'column-stacked', 'column-positive-negative'].indexOf(settings.type) !== -1) {
+    'column-grouped', 'column-stacked', 'column-positive-negative', 'positive-negative'].indexOf(settings.type) !== -1) {
     charts.clickedLegend = true;
     selector.dispatch('click');
   }
 
   if (elem.selectionObj) {
-    charts.selectElement(d3.select(elem.selectionObj.nodes()[idx]), elem.selectionInverse, elem.data); // eslint-disable-line
+    charts.selectElement(d3.select(elem.selectionObj.nodes()[idx]), elem.selectionInverse, elem.data, undefined, settings.dataset, noTrigger); // eslint-disable-line
   }
 };
 
 // The selected array for this instance.
 charts.selected = [];
+
+/**
+ * Delete all `selected` keys/value from given dataset.
+ * @private
+ * @param  {array} dataset  The data object
+ * @returns  {void}
+ */
+charts.clearSelected = function (dataset) {
+  if (dataset) {
+    const deleteSelected = (obj) => {
+      if (Object.prototype.hasOwnProperty.call(obj, 'selected')) {
+        delete obj.selected;
+      }
+    };
+    const clear = (ds) => {
+      if (Array.isArray(ds)) {
+        ds.forEach((node) => {
+          deleteSelected(node);
+          if (node.data) {
+            clear(node.data);
+          }
+        });
+      }
+      deleteSelected(ds);
+    };
+    clear(dataset);
+  }
+};
 
 /**
  * Select the element and fire the event, make the inverse selector opace.
@@ -495,9 +525,15 @@ charts.selected = [];
  * @param  {object} inverse The opposite selection.
  * @param  {array} data  The data object
  * @param  {object} container  The DOM object
+ * @param  {array} dataset  The dataset object
+ * @param  {boolean} noTrigger  if true will not trigger
  */
-charts.selectElement = function (element, inverse, data, container) {
+charts.selectElement = function (element, inverse, data, container, dataset, noTrigger) {
   const isSelected = element.node() && element.classed('is-selected');
+  charts.clearSelected(dataset);
+  if (!isSelected) {
+    data.selected = true;
+  }
   const triggerData = [{ elem: element.nodes(), data: (!isSelected ? data : {}) }];
 
   inverse.classed('is-selected', false)
@@ -509,7 +545,9 @@ charts.selectElement = function (element, inverse, data, container) {
   charts.selected = $.isEmptyObject(triggerData[0].data) ? [] : triggerData;
 
   // Fire Events
-  $(container).triggerHandler('selected', [triggerData]);
+  if (!noTrigger) {
+    $(container).triggerHandler('selected', [triggerData]);
+  }
 };
 
 /**
@@ -520,10 +558,10 @@ charts.selectElement = function (element, inverse, data, container) {
  */
 charts.setSelectedElement = function (o) {
   let dataset = o.dataset;
-  const isPositiveNegative = o.type === 'column-positive-negative';
+  const isPositiveNegative = /positive-negative/.test(o.type);
   const isBar = /^(bar|bar-stacked|bar-grouped|bar-normalized)$/.test(o.type);
   const isTypePie = o.type === 'pie' || o.type === 'donut';
-  const isTypeColumn = /^(column|column-grouped|column-stacked|column-positive-negative)$/.test(o.type);
+  const isTypeColumn = /^(column|column-grouped|column-stacked|column-positive-negative|positive-negative)$/.test(o.type);
 
   const svg = o.svg;
   const isSingle = o.isSingle;
@@ -556,9 +594,15 @@ charts.setSelectedElement = function (o) {
   pnPositiveText.style('font-weight', 'normal');
   pnNegativeText.style('font-weight', 'normal');
   svg.selectAll('.is-selected').classed('is-selected', false);
+  charts.clearSelected(o.dataset);
 
   if (isTypePie) {
     svg.selectAll('.is-not-selected').classed('is-not-selected', false);
+  }
+  if (isPositiveNegative) {
+    if (Object.prototype.hasOwnProperty.call(o.dataset[0], 'targetBarsSelected')) {
+      delete o.dataset[0].targetBarsSelected;
+    }
   }
 
   // Task make selected
@@ -570,8 +614,8 @@ charts.setSelectedElement = function (o) {
       if (isPositiveNegative) {
         if (o.isTargetBar) {
           o.svg.selectAll('.target-bar').classed('is-selected', true).style('opacity', 1);
-
           pnTargetText.style('font-weight', 'bolder');
+          o.dataset[0].targetBarsSelected = true;
         } else {
           o.svg.selectAll(isPositive ?
             '.bar.positive, .target-bar.positive' : '.bar.negative, .target-bar.negative')
@@ -583,7 +627,9 @@ charts.setSelectedElement = function (o) {
         svg.selectAll('.bar').each(function (d, i) {
           const bar = d3.select(this);
           if (bar.classed('is-selected')) {
-            selectedBars.push({ elem: bar.node(), data: (dataset ? dataset[i] : d) });
+            const bardata = dataset ? dataset[i] : d;
+            bardata.selected = true;
+            selectedBars.push({ elem: bar.node(), bardata });
           }
         });
         triggerData = selectedBars;
@@ -605,21 +651,28 @@ charts.setSelectedElement = function (o) {
             thisData = d;
           }
 
-          if (isBar) {
-            if (thisData[i][o.i]) {
-              thisData = thisData[i][o.i];
-            }
+          if (isBar && !isStacked) {
+            if (isGrouped) {
+              thisData = o.dataset[i].data[o.i];
+            } else {
+              if (thisData[i][o.i]) {
+                thisData = thisData[i][o.i];
+              }
 
-            if (thisData[o.i] && thisData[o.i][i]) {
-              thisData = thisData[o.i][i];
-            }
+              if (thisData[o.i] && thisData[o.i][i]) {
+                thisData = thisData[o.i][i];
+              }
 
-            if (thisData[i] && thisData[i][o.i]) {
-              thisData = thisData[i][o.i];
+              if (thisData[i] && thisData[i][o.i]) {
+                thisData = thisData[i][o.i];
+              }
             }
           } else if (isStacked && !isSingle) {
             if (thisData[thisGroupId] && thisData[thisGroupId].data[i]) {
               thisData = thisData[thisGroupId].data[i];
+            }
+            if (isBar) {
+              o.dataset[thisGroupId].selected = true;
             }
           } else {
             if (thisData[i].data[o.i]) {
@@ -635,6 +688,7 @@ charts.setSelectedElement = function (o) {
             }
           }
 
+          thisData.selected = true;
           selectedBars.push({ elem: bar.node(), data: thisData });
         });
         triggerData = selectedBars;
@@ -669,7 +723,9 @@ charts.setSelectedElement = function (o) {
         thisGroup.selectAll('.bar').each(function (d, i) {
           const bar = d3.select(this);
           if (bar.classed('is-selected')) {
-            selectedBars.push({ elem: bar.node(), data: (dataset ? dataset[i] : d) });
+            const data = dataset ? dataset[i] : d;
+            data.selected = true;
+            selectedBars.push({ elem: bar.node(), data });
           }
         });
         if (isGrouped) {
@@ -694,8 +750,10 @@ charts.setSelectedElement = function (o) {
         const bar = d3.select(this);
         let data = d;
         if (dataset) {
-          data = isBar && isStacked ? dataset[i][o.i] : dataset[i].data[o.i];
+          data = isBar && isStacked && typeof dataset[i][o.i] !== 'undefined' ?
+            dataset[i][o.i] : dataset[i].data[o.i];
         }
+        data.selected = true;
         selectedBars.push({ elem: bar.node(), data });
       });
       triggerData = selectedBars;
@@ -708,7 +766,7 @@ charts.setSelectedElement = function (o) {
 
       const thisArcData = dataset && dataset[0] && dataset[0].data ?  //eslint-disable-line
         dataset[0].data[o.i] : (o.d ? o.d.data : o.d);  //eslint-disable-line
-
+      thisArcData.selected = true;
       selector.classed('is-selected', true)
         .classed('is-not-selected', false)
         .attr('transform', 'scale(1.025, 1.025)');